Reformat Routes files
