+++ context/app/static/js/components/Routes/Route/Route.jsx
@@ -1,17 +1,8 @@
 import React, { Suspense } from 'react';
 import PropTypes from 'prop-types';
-import styled from 'styled-components';
-import Container from '@material-ui/core/Container';
 
-import RouteLoader from './RouteLoader';
-
-const StyledContainer = styled(Container)`
-  margin-top: ${(props) => props.theme.spacing(2)}px;
-`;
-
-const MainWrapper = styled.div`
-  flex-grow: 1;
-`;
+import RouteLoader from '../RouteLoader';
+import { StyledContainer, MainWrapper } from './style';
 
 function Route({ children, disableWidthConstraint }) {
   const constrainWidthProps = disableWidthConstraint ? { maxWidth: false, disableGutters: true } : { maxWidth: 'lg' };
