+++ context/app/static/js/components/Routes/Route/index.js
@@ -1,0 +1,3 @@
+import Route from './Route';
+
+export default Route;
