+++ context/app/static/js/components/Routes/RouteLoader/index.js
@@ -1,0 +1,3 @@
+import RouteLoader from './RouteLoader';
+
+export default RouteLoader;
