+++ context/app/static/js/components/Routes/RouteLoader/style.js
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+const LoadingWrapper = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 100vh;
+`;
+
+export { LoadingWrapper };
